+++ wasm-util/detectOther.ts
@@ -1,6 +1,6 @@
 // Copyright 2019-2022 @polkadot/wasm-util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot@0.0.0-8/wasm-util/packageInfo.ts';
+import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot/wasm-util/packageInfo.ts';
 
 export default [utilInfo];