+++ util/json.d.ts
@@ -3,7 +3,7 @@
 
 // Be able to import json in TS
 // https://stackoverflow.com/questions/49996456/importing-json-file-in-typescript
-declare module 'https://esm.sh/*.json' {
+declare module '*.json' {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const value: any;