+++ hw-ledger/detectOther.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2022 @polkadot/hw-ledger authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { packageInfo as transportInfo } from 'https://deno.land/x/polkadot@0.0.0-9/hw-ledger-transports/packageInfo.ts';
-import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot@0.0.0-9/util/packageInfo.ts';
+import { packageInfo as transportInfo } from 'https://deno.land/x/polkadot/hw-ledger-transports/packageInfo.ts';
+import { packageInfo as utilInfo } from 'https://deno.land/x/polkadot/util/packageInfo.ts';
 
 export default [transportInfo, utilInfo];