+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, hw-ledger-transports, keyring, networks, util, util-crypto, x-bigint, x-fetch, x-global, x-randomvalues, x-textdecoder, x-textencoder, x-ws 9.6.2-2-x
 - phishing 0.15.1
 - hw-ledger, hw-ledger-transports, keyring, networks, util, util-crypto, x-bigint, x-fetch, x-global, x-randomvalues, x-textdecoder, x-textencoder, x-ws 9.6.1
 - wasm-bridge, wasm-crypto, wasm-crypto-asmjs, wasm-crypto-init, wasm-crypto-wasm, wasm-util 6.1.5