+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- phishing 0.15.2
 - hw-ledger, hw-ledger-transports, keyring, networks, util, util-crypto, x-bigint, x-fetch, x-global, x-randomvalues, x-textdecoder, x-textencoder, x-ws 9.6.2
 
 ## 0.0.0-9 Jun 23, 2022